requirements.txt
Django==1.11.11
django-summernote==0.8.8.6
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
PyJWT==1.6.1
pytz==2018.3
